requirements.txt
# Core data
pandas>=2.2.2
numpy>=1.26.4
duckdb>=0.10.3

# Machine learning
scikit-learn>=1.5.0
joblib>=1.4.2

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.1
python-dotenv>=1.0.1

# Streamlit dashboard
streamlit>=1.35.0
plotly>=5.22.0

# GenAI
openai>=1.30.1

# Utilities
pyyaml>=6.0.1

# Testing
pytest>=8.2.1
pytest-asyncio>=0.23.7
httpx>=0.27.0